+++ app/javascript/form_inline.js
@@ -1,7 +1,40 @@
-import { FormInline } from './utilities/FormInline.js'
+document.addEventListener('turbolinks:load', function() {
+  let controls = document.querySelectorAll('.form-inline-link')
 
-document.addEventListener('turbolinks:load', function () {
-  const editTest = ".form-inline-link"
-  const edit = document.querySelectorAll(editTest)
-  if(edit) new FormInline(edit)
+  if (controls.length) {
+    for (let i = 0; i < controls.length; i++) {
+      controls[i].addEventListener('click', formInlineLinkHandler)
+    }
+  }
+
+  let errors = document.querySelector('.resource-errors')
+
+    if(errors) {
+      let resourceId = errors.dataset.resourceId
+      formInlineHandler(resourceId)
+    }
 })
+
+function formInlineLinkHandler(event) {
+  event.preventDefault()
+
+  let testId = this.dataset.testId
+  formInlineHandler(testId)
+}
+
+function formInlineHandler(testId) {
+  let link = document.querySelector('.form-inline-link[data-test-id="'+ testId +'"]')
+  let testTitle = document.querySelector('.test-title[data-test-id="'+ testId +'"]')
+  let formInLine = document.querySelector('.form-inline[data-test-id="'+ testId +'"]')
+
+  if(formInLine.classList.contains('hide')) {
+    testTitle.classList.add('hide')
+    formInLine.classList.remove('hide')
+    link.textContent = 'Выйти'
+  }
+  else {
+    testTitle.classList.remove('hide')
+    formInLine.classList.add('hide')
+    link.textContent = 'Редактировать тест'
+  }
+}
